pyproject.toml
[project]
name = "finances"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "Flask>=3.0",
    "PyYAML>=6.0",
    "tabulate>=0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=7.0",
    "ruff>=0.8",
    "jsonschema>=4.0",
    "playwright>=1.40",
    "pytest-playwright>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "e2e: end-to-end browser tests (deselect with '-m not e2e')",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
ignore = ["E501"]
